refactor: move Diving and Moim to types
