+++ components/Search/MoimItem.tsx
@@ -1,7 +1,7 @@
 import { View, Text, Image, TouchableOpacity } from 'react-native';
 import TagBox from './styles/TagBox';
 import { router } from 'expo-router';
-import { Moim } from 'api/moim/moim';
+import { Moim } from 'types/Gatherings';
 import { tokens, images, icons } from 'constants/';
 
 interface MoimItemProps {
